\name{agents-package}
\alias{agents-package}
\alias{agents}
\docType{package}
\title{
\packageTitle{agents}
}
\description{
\packageDescription{agents}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{agents}
\packageIndices{agents}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{agents}
Maintainer: \packageMaintainer{agents}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
