% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agents.R
\docType{package}
\name{agents}
\alias{agents}
\title{agents package}
\description{
Sequential Monte Carlo algorithms for agent-based models
}
\details{
This R package implements sequential Monte Carlo algorithms for SIS and SIR agent-based models.
}
\author{
Phyllis Ju <nianqiao@purdue.edu>
}
\keyword{package}
