% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config.R
\name{check_model_config}
\alias{check_model_config}
\title{check  model configurations}
\usage{
check_model_config(model_config)
}
\arguments{
\item{model_config}{a list that must contain
\itemize{
\item N, population size;
\item network_type, full or sparse;
\item for sparse network, must give a matrix `neighbors`, whose rows look like (1,2,4,-1,-1).;
\item parameters alpha0,lambda,gamma, and rho.
}}
}
\description{
this function checks if model_config is eligible as input for simulations, particle filters and gibbs samplers.
}
