% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config.R
\name{check_particle_config}
\alias{check_particle_config}
\title{check particle configurations}
\usage{
check_particle_config(particle_config)
}
\arguments{
\item{particle_config}{a list that contains that must contain: TODO}
}
\description{
this function checks if particle_config is eligible as input for particle filters
}
