% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distance_metropolis_idchecking.R
\name{compute_distance_metropolis_to_condbern}
\alias{compute_distance_metropolis_to_condbern}
\title{The Metropolis sampler for Conditional Bernoulli distribution}
\usage{
compute_distance_metropolis_to_condbern(
  alpha,
  sum_x,
  num_mcmc,
  num_repeat,
  distance = "tv"
)
}
\arguments{
\item{alpha}{vector of probabilities for the Poisson-Binomial distribution}

\item{sum_x}{sum of all entries}

\item{num_mcmc}{number of MCMC iterations to run before coupling}

\item{num_repeat}{number of repetitions to get an estimate of the upper bound}

\item{distance}{can be 'tv' or 'wasserstein'}
}
\value{
a vector of distance upper bounds for different number of iterations
}
\description{
compute an upper bound for the 2-Wasserstein distance 
between Metropolis sampler in metropolis_condbern.R and the Conditional-Bernoulli distribution
}
