% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_tv_poisbinom_translatedpois.R
\name{compute_tv_poisbinom_translatedpois}
\alias{compute_tv_poisbinom_translatedpois}
\title{Compute the TV distance between Poisson-Binomial distribution and its translated Poisson approximation}
\usage{
compute_tv_poisbinom_translatedpois(alpha)
}
\arguments{
\item{alpha}{vector of probabilities for the Poisson-Binomial distribution}
}
\value{
total variation distance
}
\description{
Compute the TV distance between Poisson-Binomial distribution and its translated Poisson approximation
}
