% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couple_condbern_kernel.R
\name{couple_condbern_kernel}
\alias{couple_condbern_kernel}
\title{a coupling kernel that leaves conditional-bernoullis invariant}
\usage{
couple_condbern_kernel(xx1, xx2, sum_x1, sum_x2, alpha1, alpha2)
}
\arguments{
\item{xx1}{current state of chain1}

\item{xx2}{current state of chain2}

\item{sum_x1}{sum of xx1}

\item{sum_x2}{sum of xx2}

\item{alpha1}{vector of probabilities for chain1}

\item{alpha2}{vector of probabilities for chain2}
}
\value{
a matrix row 1 is new xx1 and row 2 is new xx2
}
\description{
discription of the algorithm in page 21 of Jacob et al. (2017).
for one chain, it samples index of a 0 entry in X uniformly and index of a 1 entry uniformly
it proposes a switch of the 1 and 0 and accepts the swap with probability ratio p(proposal) / p(current)
}
