% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/couple_metropolis_to_condbern.R
\name{couple_metropolis_to_condbern}
\alias{couple_metropolis_to_condbern}
\title{Coupled Metropolis algorithm for the Conditional-Bernoulli distribution}
\usage{
couple_metropolis_to_condbern(
  alpha,
  sum_x,
  num_mcmc,
  compute_sse = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{alpha}{vector of probabilities for the Poisson-Binomial distribution}

\item{sum_x}{sum of x}

\item{num_mcmc}{number of MCMC iterations; if num.mcmc = Inf, then run the algorithm until the two chains couple}

\item{compute_sse}{binary, default is FALSE; if compute_sse = TRUE, then returns the sum-squared-error distance from chain1 to chain2}

\item{verbose}{binary, whether to print messages}
}
\value{
a list that contains 
\itemize{
\item meeting time 
\item sum-squared-error distance from chain1 to chain2
}
}
\description{
A coupling of two metropolis samplers targetting the same Conditional-Bernoulli distribution
Chain1 starts from the stationary distribution and chain2 is initialized uniformly from the support
}
