% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpoibin.R
\name{dpoibin_exact}
\alias{dpoibin_exact}
\title{Poisson binomial probability mass function}
\usage{
dpoibin_exact(eval, alpha, log_scale = TRUE)
}
\arguments{
\item{eval}{a vector of evaluation points}

\item{alpha}{a vector or matrix of probabilities}

\item{log_scale}{logical specifying if probabilities are evaluated on the logarithmic scale}
}
\value{
a vector or matrix of probabilities
}
\description{
Evaluates Poisson binomial probability mass function exactly
}
