% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsumbin.R
\name{dsumbin}
\alias{dsumbin}
\title{Probability mass function of the sum of two independent binomial random variables}
\usage{
dsumbin(eval, n1, p1, n2, p2, log_scale = TRUE)
}
\arguments{
\item{eval}{a vector of evaluation points}

\item{n1}{number of trials of first binomial random variable}

\item{p1}{probability of success of first binomial random variable}

\item{n2}{number of trials of second binomial random variable}

\item{p2}{probability of success of second binomial random variable}

\item{log_scale}{logical specifying if probabilities are evaluated on the logarithmic scale}
}
\value{
a vector or matrix of probabilities
}
\description{
Evaluates probability mass function of the sum of two independent binomial random variables
}
