% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtranspoi.R
\name{dtranspoi}
\alias{dtranspoi}
\title{Probability mass function of a translated Poisson distribution}
\usage{
dtranspoi(eval, mu, sigmasq, end_support, log_scale = TRUE)
}
\arguments{
\item{eval}{a vector of evaluation points}

\item{mu}{a vector of means}

\item{sigmasq}{a vector of variances}

\item{end_support}{upper end of support}

\item{log_scale}{logical specifying if probabilities are evaluated on the logarithmic scale}
}
\value{
a vector or matrix of probabilities
}
\description{
Evaluates probability mass function of a translated Poisson distribution
}
