% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dtranspoi.R
\name{dtranspoi_poibin_approx}
\alias{dtranspoi_poibin_approx}
\title{Probability mass function of a translated Poisson approximation of a Poisson binomial distribution}
\usage{
dtranspoi_poibin_approx(eval, alpha, log_scale = TRUE)
}
\arguments{
\item{eval}{a vector of evaluation points}

\item{alpha}{a vector or matrix of probabilities}

\item{log_scale}{logical specifying if probabilities are evaluated on the logarithmic scale}
}
\value{
a vector or matrix of probabilities
}
\description{
Evaluates probability mass function of a translated Poisson approximation of a Poisson binomial distribution
}
