% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_distance_metropolis_idchecking.R
\name{get_meetingtime_metropolis_idchecking}
\alias{get_meetingtime_metropolis_idchecking}
\title{The Metropolis sampler for Conditional Bernoulli distribution}
\usage{
get_meetingtime_metropolis_idchecking(
  alpha,
  sum_x,
  num_repeat,
  max_iterations = Inf
)
}
\arguments{
\item{alpha}{vector of probabilities}

\item{sum_x}{Conditional-Bernoulli constraint, sum of all entries}

\item{num_repeat}{number of repetitions}

\item{max_iterations}{maximum number of iterations, at which to interrup the while loop; Inf by default}
}
\value{
Meeting times stored in a vector
}
\description{
meeting time between Metropolis sampler in gibbs_condbern.R and the Conditional-Bernoulli distribution
}
