% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rates_from_features.R
\name{get_rates_from_features}
\alias{get_rates_from_features}
\title{Parameters for the SIS and SIR process}
\usage{
get_rates_from_features(beta, features)
}
\arguments{
\item{beta}{a vector of coefficients. Its length equals the number of features.}

\item{features}{N by num_features matrix}
}
\value{

}
\description{
Parameters for the SIS and SIR process
}
