% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idcheck_sampling.R
\name{idcheck_sampling}
\alias{idcheck_sampling}
\title{ID-check sampling}
\usage{
idcheck_sampling(sum_x, alpha, config = list(num_sample = 1))
}
\arguments{
\item{sum_x}{sum of the samples}

\item{alpha}{vector of probabilities}

\item{config}{a list containing
\itemize{
\item num_sample : number of samples from conditional-bernoulli; default is 1
}}
}
\value{
a matrix or a vector containing the Conditional Bernoulli samples
}
\description{
obtain a sample from the conditional Bernoulli distribution 
uses the recursive method 2 from Chen and Liu (1997)
}
