% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdpoisbinom.R
\name{logdpoisbinom}
\alias{logdpoisbinom}
\title{Poisson Binomial distribution}
\usage{
logdpoisbinom(alpha)
}
\arguments{
\item{alpha}{vector of N probabilities (numbers between 0 and 1).}
}
\value{
the entire density function in a vector of length N+1.
}
\description{
Density function for Poisson binomial distribution with parameters alpha, 
which is the distribution of the sum of Bernoulli random variables, each with a different
probability \eqn{P(X_i = 1) = \alpha_i}.
}
