% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdsumbin.R
\name{logdsumbin}
\alias{logdsumbin}
\title{Sum of two independent Binomials}
\usage{
logdsumbin(i_current, lambda_bar, gamma_bar, N)
}
\arguments{
\item{lambda_bar}{average infection rate}

\item{gamma_bar}{average recovery rate}

\item{it}{number of infections}
}
\value{
log transition densities p(i(t+1) = 0 : N | it, lambda_bar, gamma_bar);
}
\description{
computes the log density of SumBinomial distribition, parametrized by lambda_bar, gamma_bar and it
}
