% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdtranspoisson.R
\name{logdtranspoisson}
\alias{logdtranspoisson}
\title{translated Poisson distribution}
\usage{
logdtranspoisson(xx, mu, sigmasq)
}
\arguments{
\item{xx}{vector of integers}

\item{mu}{mean}

\item{sigmasq}{variance}
}
\value{
log densities
}
\description{
density approximation of the Poission binomial distribution in log scale
}
