% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdtranspoisson_approx.R
\name{logdtranspoisson_approx}
\alias{logdtranspoisson_approx}
\title{translated Poisson approximation}
\usage{
logdtranspoisson_approx(alpha, y = NULL)
}
\arguments{
\item{alpha}{vector of Bernoulli parameters}

\item{xx}{default is NULL, when x == NULL, output is the entire density function}
}
\value{
approximate log densities.
}
\description{
density approximation of the Poission binomial distribution in log scale. \\
NOTE: the densities do not sum to 1.
}
