% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsumexp.R
\name{lw.normalize}
\alias{lw.normalize}
\alias{lw.logsum}
\alias{lw.logmean}
\alias{lw.logsd}
\title{Normalize log weights using the log-sum-exp trick.}
\usage{
lw.normalize(lw)

lw.logsum(lw)

lw.logmean(lw)

lw.logsd(lw)
}
\arguments{
\item{lw}{log weights}
}
\value{
\code{lw.normalize} returns normalized weights.

\code{lw.normalize} computes the log normalizing constant

\code{lw.logmean} computes log of mean of the vector \eqn{exp(lw)}.

\code{lw.logsd} computes log of standard deviation of the vector \eqn{exp(lw)}.
}
\description{
Normalize log weights using the log-sum-exp trick.
}
