% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metropolis_condbern.R
\name{metropolis_condbern}
\alias{metropolis_condbern}
\title{Metropolis algorithm targeting the Conditional Bernoulli distribution}
\usage{
metropolis_condbern(sum_x, alpha, num_sample = 10, save_history = FALSE)
}
\arguments{
\item{sum_x}{sum of the samples}

\item{alpha}{vector of probabilities, length N}

\item{config}{list of configurations, save_history, num_sample}
}
\value{
either xx (a vector of size N) or  xx_matrix (a binary matrix of size num_samples by N)
}
\description{
This is an approximation of idcheck_sampling. Given the current state x, 
it proposes to switch a randomly chosen 1 with a randomly chosen 0.
}
