% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{multinomial_resampling}
\alias{multinomial_resampling}
\title{Multinomial resampling scheme}
\usage{
multinomial_resampling(weights, ndraws)
}
\arguments{
\item{weights}{a vector of normalized weights}

\item{ndraws}{integer specifying number of ancestor indexes required}
}
\value{
a vector of ancestor indexes
}
\description{
Implements the multinomial resampling scheme
}
