% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylast.R
\name{mylast}
\alias{mylast}
\title{Access the last element in a vector}
\usage{
mylast(x)
}
\arguments{
\item{x}{vector}
}
\value{
last element in the vector x
}
\description{
Access the last element in a vector
}
