% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_fully_connected.R
\name{network_fully_connected}
\alias{network_fully_connected}
\title{Network structure}
\usage{
network_fully_connected(N)
}
\arguments{
\item{N}{integer constant, the population size.}
}
\description{
generate the (normalized) adjacency matrix of a fully connected network
}
