% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network_sample_smallworld.R
\name{network_sample_smallworld}
\alias{network_sample_smallworld}
\title{Small world graph}
\usage{
network_sample_smallworld(N, dim = 1, nei = 1, p = 0.3)
}
\arguments{
\item{N}{integer constant, the population size.}

\item{dim}{integer constant, dimension of the starting lattice;}

\item{nei}{integer constant, the neighborhood within which the vertices of the lattice will be connected.}

\item{p}{real constant between zero and one, the rewiring probability.}
}
\description{
sample a small world graph and return the (normalized) adjacency matrix
}
