% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcat.R
\name{rcat}
\alias{rcat}
\title{Random sample from categorical distribution}
\usage{
rcat(support, prob, num_samples = 1)
}
\arguments{
\item{support}{a vector specifying the support}

\item{prob}{a vector or matrix of probabilities}

\item{num_samples}{number of samples required (default is one)}
}
\value{
a vector of samples
}
\description{
Generate a sample from categorical distribution
}
