% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcondber.R
\name{rcondber_exact}
\alias{rcondber_exact}
\title{Random sample from conditional Bernoulli distribution}
\usage{
rcondber_exact(conditioned_sum, alpha)
}
\arguments{
\item{conditioned_sum}{a vector of conditioned sums}

\item{alpha}{a vector or matrix of probabilities}
}
\value{
a logical matrix
}
\description{
Generate a sample from conditional Bernoulli distribution
}
