% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcondber.R
\name{rcondber_mcmc}
\alias{rcondber_mcmc}
\title{Random sample from conditional Bernoulli distribution using Markov chain Monte Carlo}
\usage{
rcondber_mcmc(conditioned_sum, alpha, mcmc_iterations)
}
\arguments{
\item{conditioned_sum}{a vector of conditioned sums}

\item{alpha}{a vector or matrix of probabilities}

\item{mcmc_iterations}{number of Markov chain Monte Carlo iterations}
}
\value{
a logical matrix
}
\description{
Generate a sample from conditional Bernoulli distribution using Markov chain Monte Carlo
}
