% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seroprevalence_survey.R
\name{seroprevalence_survey_model}
\alias{seroprevalence_survey_model}
\title{Seroprevalence survey model}
\usage{
seroprevalence_survey_model(sample_size, population_size)
}
\arguments{
\item{sample_size}{sample size of survey}

\item{population_size}{population size}
}
\value{
a list containing: 
\itemize{
\item 'rmeasurement': function to sample from observation model
\item 'dmeasurement: function evaluating observation density
\item 'infected_support': function returning support of number of infected given observation  
}
}
\description{
Returns objects defining seroprevalence survey model
}
