% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_apf.R
\name{sir_apf}
\alias{sir_apf}
\title{Auxilliary particle filter for the agent-based hidden SIR process.}
\usage{
sir_apf(y, model_config, particle_config)
}
\arguments{
\item{y}{a list of observations, length (T+1).}

\item{model_config}{a list containing model parameters, must pass the test of check_model_config;}

\item{particle_config}{a list specifying the particle filter, must pass the test of check_particle_config;}
}
\value{
the particle filter outcome, must contain estimate of log marginal likelihood and effective sample size.
}
\description{
unbiased approximation of the marginal likelihood. For this function, the observation density is \deqn{y_t \mid x_t \sim \textrm{Bin}(I(x_t),\rho).}
}
