% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_backward_information_filter_si.R
\name{sir_backward_information_filter_si}
\alias{sir_backward_information_filter_si}
\title{Controlled SMC for SIR model}
\usage{
sir_backward_information_filter_si(y, model_config)
}
\arguments{
\item{y}{population observations}

\item{model_config}{a list containing model parameters and network structure}
}
\value{
policy approximate dynamic programming matrix pre-calculated to run an smc filter. This is an array of size [N + 1, N + 1, T + 1]
}
\description{
computes the backward information filter as a function of s and i. This function will replace the current sir_backward_information_filter_sumbin once completed.
}
