% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_backward_information_filter_sumbin_dev.R
\name{sir_backward_information_filter_sumbin_dev}
\alias{sir_backward_information_filter_sumbin_dev}
\title{Controlled SMC for SIR model}
\usage{
sir_backward_information_filter_sumbin_dev(y, model_config, index2i, index2r)
}
\arguments{
\item{y}{population observations}

\item{beta}{infection rate}

\item{gamma}{recovery rate}

\item{rho}{reporting rate}
}
\value{
policy approximate dynamic programming matrix pre-calculated to run an smc filter. This is an array of size [N + 1, N + 1, T + 1]
}
\description{
computes the backward information filter
}
