% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_csmc.R
\name{sir_csmc}
\alias{sir_csmc}
\title{controlled SMC for SIR model}
\usage{
sir_csmc(y, model_config, particle_config, policy = NULL)
}
\arguments{
\item{y}{, observations, length (T+1)}
}
\value{
A list containing
}
\description{
marginal likelihood estimations and samples from the posterior distribution of agent states
}
