% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_csmc_si.R
\name{sir_csmc_si}
\alias{sir_csmc_si}
\title{controlled SMC for SIR model}
\usage{
sir_csmc_si(y, model_config, particle_config, logpolicy = NULL)
}
\arguments{
\item{y}{, observations, length (T+1)}

\item{model_config}{a list containing model parameters, must pass the test of check_model_config;}

\item{particle_config}{a list specifying the particle filter, must pass the test of check_particle_config;}
}
\value{
the particle filter outcome, must contain estimate of log marginal likelihood and effective sample size.
}
\description{
marginal likelihood estimations and samples from the posterior distribution of agent states. This function is the development version of sir_csmc. 
Its policy will be a function of (st,it).
}
