% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_get_alpha.R
\name{sir_get_alpha_sir}
\alias{sir_get_alpha_sir}
\alias{sir_get_alpha_i}
\alias{sir_get_alpha_plusone}
\title{SIR model transition kernels}
\usage{
sir_get_alpha_sir(agent_states, model_config)

sir_get_alpha_i(agent_states, model_config)

sir_get_alpha_plusone(agent_states, model_config)
}
\arguments{
\item{agent_states}{a vector containing agent states at time \eqn{t}.}

\item{model_config}{a list containing model parameters, which must pass the function \code{check_model_config()}.}
}
\value{
\code{sir_get_alpha_sir(agent_states,model_config)} returns a matrix with three rows : alpha_s, alpha_i and alpha_r.

\code{sir_get_alpha_i(agent_states, model_config)} returns alpha_i as a vector.

\code{sir_get_alpha_plusone(agent_states, model_config)} returns a vector, whose entries are the probabilities that p(x[t+1,k] = x[t,k] + 1);
}
\description{
The alpha function for SIR process gives the transition probabilities from \eqn{x[t,n]} to \eqn{x[t+1,n]} given \eqn{x[t]}.
}
