% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_get_state_space.R
\name{sir_get_state_space}
\alias{sir_get_state_space}
\title{State space for SIR model}
\usage{
sir_get_state_space(N)
}
\arguments{
\item{N}{population size}
}
\value{
the space, as a 3**N by N matrix
}
\description{
create the space \{0,1,2\}^N
}
\examples{
sir_get_state_space(3)
}
