% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_kernel.R
\name{sir_kernel}
\alias{sir_kernel}
\title{transition kernel for the SIR process}
\usage{
sir_kernel(xxprev, model_config)
}
\arguments{
\item{xxprev}{which is $X_{t-1}$, length N vector presenting the current states}

\item{model_config}{a list containing model paramters}
}
\value{
xx which is $X_t$
}
\description{
This is the one step transition kernel
}
