% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_kernel_twisted_i.R
\name{sir_kernel_twisted_i}
\alias{sir_kernel_twisted_i}
\title{Twisted SIR kernels}
\usage{
sir_kernel_twisted_i(xxprev, it, model_config)
}
\arguments{
\item{xxprev}{agent states at the previous time;}

\item{it}{infection counts at the current time;}

\item{model_config}{a list containing model parameters, must pass the test of check_model_config;}
}
\value{
a sample of xxprev, vector of length N.
}
\description{
samples from xt given it and x(t-1). This function is used in sir_apf.R.
}
