% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_kernel_twisted_ir.R
\name{sir_kernel_twisted_ir}
\alias{sir_kernel_twisted_ir}
\title{Twisted SIR kernels}
\usage{
sir_kernel_twisted_ir(xxprev, it, rt, model_config)
}
\arguments{
\item{xxprev}{agent states at the previous time;}

\item{it}{infection counts at the current time;}

\item{rt}{recovery counts at the current time;}

\item{model_config}{a list containing model parameters, must pass the test of check_model_config;}
}
\value{
a sample of xxprev, vector of length N
}
\description{
Function to sample from xt given it, rt and x(t-1), used in sir_csmc.R.
}
