% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_log_transition_density.R
\name{sir_log_transition_density}
\alias{sir_log_transition_density}
\title{Markov transition kernel for SIR model}
\usage{
sir_log_transition_density(xx_diff, alpha_plusone)
}
\arguments{
\item{xx_diff}{binary vector, equals x[t+1] - x[t]}

\item{alpha_plueone}{probability of transitioning to the next state from x[t]}
}
\description{
computes the log transition density f(x[t+1] | x[t])
}
