% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_logdpoismulti_given_xxprev.R
\name{sir_logdpoismulti_given_xxprev}
\alias{sir_logdpoismulti_given_xxprev}
\title{SIR model}
\usage{
sir_logdpoismulti_given_xxprev(xxprev, model_config)
}
\arguments{
\item{xxprev}{\eqn{x[t-1]}}

\item{model_config}{a list containing parameters and network structure}
}
\value{
the matrix containing log densities of \eqn{p(st,it | x[t-1])}.
}
\description{
compute the transition probability from xx[t-1] to the aggregated state (st, it). 
NOTE: if network is full, then we can use \code{sir_csmc_update_f_matrix} instead.
}
