% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_loglikelihood_complete.R
\name{sir_loglikelihood_complete}
\alias{sir_loglikelihood_complete}
\title{SIR model likelihood}
\usage{
sir_loglikelihood_complete(xx, y, model_config)
}
\arguments{
\item{xx}{agent states}

\item{y}{population observations}

\item{model_config}{a list containinng parameters, features, and network structure}
}
\value{
log likelihood
}
\description{
p(x,y) for the SIR model
}
