% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_simulate.R
\name{sir_simulate}
\alias{sir_simulate}
\title{SIR process}
\usage{
sir_simulate(days, model_config, dt = 1)
}
\arguments{
\item{days}{length of simulation}

\item{model_config}{a list containing model parameters, passing the check_model_config function.}

\item{dt}{increment in time}
}
\value{
a list containing
\itemize{
\item agent_states : a binary matrix containing the hidden states, size is N by (days + 1)
\item y : a vector of reports, length is days + 1
}
}
\description{
Simulate SIR process as described in Section 4.1
}
