% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_xx_gibbs_blocked_kernel.R
\name{sir_xx_gibbs_blocked_kernel}
\alias{sir_xx_gibbs_blocked_kernel}
\title{SIR blocked gibbs}
\usage{
sir_xx_gibbs_blocked_kernel(
  xx,
  block_members,
  y,
  model_config,
  all_xx_block = NULL
)
}
\arguments{
\item{xx}{a matrix of size model_config$N by length(y), can be initialized from sis_init_hidden_popobservation.R}

\item{block_members}{a list of agent ids, length should not exceed 5}

\item{model_config}{a list that must contain}

\item{all_xx_block}{default is NULL. This matrix represents the set {1,2,3}^B.
\itemize{
\item N : population size
\item rho : reporting rate
\item alpha0 : vector of probabilities
\item state_to_alpha: function to compute alpha given agent states at a previous time 
}}

\item{time_step}{0 <=t <= T}
}
\value{
xx
}
\description{
update the agent states for the SIR model given observations such that the complete likelihood is not zero.
}
