% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_xx_gibbs_singlesite_kernel.R
\name{sir_xx_gibbs_singlesite_kernel}
\alias{sir_xx_gibbs_singlesite_kernel}
\title{Gibbs sampler for SIR model}
\usage{
sir_xx_gibbs_singlesite_kernel(xx, y, model_config)
}
\arguments{
\item{xx}{current agent states}

\item{y}{observations}

\item{model_config}{a list containing paraemters, features and network structure}
}
\value{
xx
}
\description{
Gibbs sampler for SIR model
}
