% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_xx_gibbs_swap_kernel.R
\name{sir_xx_gibbs_swap_kernel}
\alias{sir_xx_gibbs_swap_kernel}
\title{Gibbs sampler for SIR model}
\usage{
sir_xx_gibbs_swap_kernel(xx, ij, y, model_config)
}
\arguments{
\item{xx}{current state x}

\item{ij}{a vector [i,j]}

\item{y}{observations}
}
\description{
given indices i and j , the proposed new state $Z$ is such that $Z^i \leftarrow X^j$ and $Z^j \leftarrow X^i$ and $Z^k \leftarrow X^k$ for all $k\not= i,j$. We accept the proposed state with probability \deqn{\min\{1, \frac{p_{\theta}(Z,y)}{p_{\theta}(X,y)}\}.}
}
