% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_xx_initialize.R
\name{sir_xx_initialize}
\alias{sir_xx_initialize}
\title{Gibbs sampler for SIR model}
\usage{
sir_xx_initialize(y, model_config)
}
\arguments{
\item{y}{population observations}

\item{model_config}{a list containinng parameters, features, and network structure}
}
\value{
agent_states
}
\description{
initialize agent states for SIR model given y such that the complete likelihood is not zero.
}
