% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_alpha_prob.R
\name{sis_alpha_prob}
\alias{sis_alpha_prob}
\title{Transition probability of SIS agent-based model}
\usage{
sis_alpha_prob(model_config, X, I)
}
\arguments{
\item{model_config}{a list containing: 
\itemize{
\item 'lambda': vector of infection rates of length N 
\item 'gamma': vector of recovery rates of length N 
\item 'network_type': character "full" or "net" specifying if network is fully connected or has a given network structure
\item 'adjacency': matrix indicating interaction of agents of dimension N x N (if model_config$network_type == "net)
\item 'degree' vector of node degree of length N (if model_config$network_type == "net)
}}

\item{X}{matrix of agent states of dimension N x P}

\item{I}{vector infected agents of length P (to avoid additional computation if this summary is needed)}
}
\value{
matrix of transition probabilities of dimension N x P
}
\description{
Computes transition probability of SIS agent-based model.
}
