% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_backward_information_filter_sumbin.R
\name{sis_backward_information_filter_sumbin}
\alias{sis_backward_information_filter_sumbin}
\title{Backward Information Filter for SIS model}
\usage{
sis_backward_information_filter_sumbin(y, model_config)
}
\arguments{
\item{y}{a sequence of observations}

\item{model_config}{a list containing parameters and network structure}
}
\value{
approximate policy
}
\description{
Approximate backward recursion based on coarse-graining the state space.
}
