% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_backward_information_filter_tp.R
\name{sis_backward_information_filter_tp}
\alias{sis_backward_information_filter_tp}
\title{Backward information filter for SIS model}
\usage{
sis_backward_information_filter_tp(y, model_config)
}
\arguments{
\item{y}{population observations}

\item{model_config}{a list that contains parameters and network structure}
}
\value{
policy approximate dynamic programming matrix pre-calculated to run an smc filter
}
\description{
This function computes performs the backward information filters steps. 
It returns a matrix of size (N + 1) by T and each column approximates observation densities of 
observing current and all the future observations given the current state i(xt).
}
