% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_bpf.R
\name{sis_bpf}
\alias{sis_bpf}
\title{Bootstrap Particle Filter for SIS model with Population Observations}
\usage{
sis_bpf(y, model_config, particle_config)
}
\arguments{
\item{y}{a vector of length (T+1)}

\item{model_config}{a list containing:
\itemize{
\item 'N': size of the population
\item 'alpha0' : initial infection probability
\item 'lambda': infection rate 
\item 'gamma': recovery rate
\item 'adjacency_matrix_b': network structure
\item 'rho': reporting rate
\item }}

\item{particle_config}{a list containing:
\itemize{
\item 'num_particles' : number of particles for the bootstrap particle filter
\item 'ess_threshold' : if effective sample size drops below the threshold, then perform a resample step. ess_threshold = 1 means resampling at every step.
\item 'save_particles': binary
\item 'clock' : binary, default to FALSE. If clock = TRUE, then we will use a stopwatch to document its Sys.time()
\item 'save_genealogy': binary
\item 'verbose': if TRUE print messages for debug
}}
}
\value{
A list containing 
\itemize{
\item 'log_incremental_likelihood' :  
\item 'log_marginal_likelihood' :
\item 'log_final_likelihood' :
\item 'particles': [N, (T+1), num_particles] array storing the particles: 
\item 'eves' : the unique ancesters of each particle in the genealogy structure, if save_genealogy = TRUE;
\item 'ess' : effective sample size at each step
\item 'runtime' : elapse time at each step, if clock = TRUE;
\item 'totaltime' : elapse time of running the whole particle filter, if clock = TRUE;
}
}
\description{
approximates marginal likelihood of the SIS model using bootstrap particle filter
NOTE: this cannot do predictions yet.
}
