% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_get_alpha.R
\name{sis_get_alpha}
\alias{sis_get_alpha}
\title{Evoluation of the SIS model}
\usage{
sis_get_alpha(agent_state, model_config)
}
\arguments{
\item{agent_state}{the state at time t}

\item{model_config}{a list of model parameters and network structure}
}
\value{
the alpha vector
}
\description{
for the sis model, given xt,computes the probability of x[t+1,k] = 1 for each k
}
