% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_get_state_space.R
\name{sis_get_state_space}
\alias{sis_get_state_space}
\title{State space of SIS model}
\usage{
sis_get_state_space(N)
}
\arguments{
\item{N}{population size}
}
\value{
the state spacespace \{0,1\}^N
}
\description{
Please only use this function for N <= 5.
}
\examples{
all_x_sis(3)
}
