% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_loglikelihood_complete.R
\name{sis_loglikelihood_complete}
\alias{sis_loglikelihood_complete}
\title{Complete Likelihood for SIS model}
\usage{
sis_loglikelihood_complete(y, agent_states, model_config)
}
\arguments{
\item{y}{a vector of observations}

\item{model_config}{a list containing parametesr and network structure, passing the check of \code{check_model_config()}.}

\item{agent_state}{binary matrix of size N by length(y)}
}
\description{
computes \eqn{log p(x, y)} when agent states and counts are observed
}
