% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_missing_apf_lookahead.R
\name{sis_missing_apf_lookahead}
\alias{sis_missing_apf_lookahead}
\title{Look-ahead approximation of auxiliary particle filter for SIS model with missing data}
\usage{
sis_missing_apf_lookahead(
  model_config,
  observations,
  conditional_expectation = "sumbin"
)
}
\arguments{
\item{model_config}{a list containing:
\itemize{
\item 'N': size of the population
\item 'lambda': vector of infection rates of length N 
\item 'gamma': vector of recovery rates of length N 
\item 'network_type': character "full" or "net" specifying if network is fully connected or has a given network structure
\item 'adjacency': matrix indicating interaction of agents of dimension N x N (if model_config$network_type == "net)
\item 'degree' vector of node degree of length N (if model_config$network_type == "net)
\item 'dmeasurement': function evaluating observation density
\item 'infected_support': function returning support of number of infected given observation  
\item 'nsteps': number of time steps
}}

\item{observations}{a list containing:
\itemize{
\item 'T': number of observations is T+1
\item 'y': vector of observations of length T+1
\item 'times': vector of length nsteps+1 indicating the observation times
}}

\item{conditional_expectation}{a character specifying how conditional expectation is computed or approximated:
\itemize{
\item 'sumbin': exact computation of conditional expectations
\item 'transpoi': translated Poisson approximation of conditional expectations
}}
}
\value{
approximate backward information filter stored as matrix of dimension N+1 x T+1
}
\description{
Look-ahead approximation of auxiliary particle filter for SIS agent-based model with missing data
}
