% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_missing_simulate.R
\name{sis_missing_simulate}
\alias{sis_missing_simulate}
\title{Simulate SIS model}
\usage{
sis_missing_simulate(model_config, observations)
}
\arguments{
\item{model_config}{a list containing:
\itemize{
\item 'N': size of the population
\item 'alpha0': initial infection probability
\item 'lambda': vector of infection rates of length N 
\item 'gamma': vector of recovery rates of length N 
\item 'network_type': character "full" or "net" specifying if network is fully connected or has a given network structure
\item 'adjacency': matrix indicating interaction of agents of dimension N x N (if model_config$network_type == "net)
\item 'degree' vector of node degree of length N (if model_config$network_type == "net)
\item 'rmeasurement': function to sample from observation model
\item 'nsteps': number of time steps
}}

\item{observations}{a list containing:
\itemize{
\item 'T': number of observations is T+1
\item 'times': vector of length nsteps+1 indicating the observation times
}}
}
\value{
a list containing: 
\itemize{
\item 'X': matrix of particle states of dimension (nsteps+1) x N 
\item 'I': vector of infected agents of length nsteps+1
\item 'y': vector of observations of length T+1
}
}
\description{
Runs a simulation of SIS agent-based model
}
