% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_simulate.R
\name{sis_simulate}
\alias{sis_simulate}
\title{SIS process}
\usage{
sis_simulate(days, model_config, dt = 1)
}
\arguments{
\item{days}{length of simulation}

\item{model_config}{a list containing
\itemize{
\item N
\item adjacency matrix
\item alpha0
\item gamma
\item beta
\item rho}}

\item{dt}{increment in time}
}
\value{
a list containing
\itemize{
\item agent_states : a binary matrix containing the hidden states, size is N by (days + 1)
\item y : a vector of reports, length is days + 1
\item 
}
}
\description{
Simulate SIS process as described in Section 3.1
}
