% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_xx_gibbs_block_by_agents_kernel.R
\name{sis_xx_gibbs_block_by_agents_kernel}
\alias{sis_xx_gibbs_block_by_agents_kernel}
\title{SIS blocked gibbs}
\usage{
sis_xx_gibbs_block_by_agents_kernel(
  xx,
  block_members,
  y,
  model_config,
  all_xx_block = NULL
)
}
\arguments{
\item{xx}{a matrix of size model_config$N by length(y), can be initialized from sis_init_hidden_popobservation.R}

\item{block_members}{a list of agent ids, length should not exceed 5}

\item{y}{observations}

\item{model_config}{a list that must contain}

\item{all_xx_block}{default is NULL, this is a matrix containing the state space {0,1}^B.}
}
\value{
xx
}
\description{
update the agent states for the SIS model given observations such that the complete likelihood is not zero.
}
