% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_xx_gibbs_initialize.R
\name{sis_xx_gibbs_initialize}
\alias{sis_xx_gibbs_initialize}
\title{Gibbs sampler for hidden agent states in the SIS model}
\usage{
sis_xx_gibbs_initialize(y, model_config)
}
\arguments{
\item{y}{a vector of population observations}

\item{model_config}{a list containing model parameters, must pass the function \code{check_model_config}.}
}
\value{
a binary matrix of agent states in the SIS model. It ensures that the complete likelihood is not zero.
}
\description{
It initializes the agent states for the SIS model given observations. This function ensures that that the complete likelihood is not zero.
}
