% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_xx_gibbs_singlesite_kernel.R
\name{sis_xx_gibbs_singlesite_kernel}
\alias{sis_xx_gibbs_singlesite_kernel}
\title{Gibbs sampler for SIS model}
\usage{
sis_xx_gibbs_singlesite_kernel(xx, y, model_config)
}
\arguments{
\item{xx}{a matrix of size model_config$N by length(y), can be initialized from sis_init_hidden_popobservation.R}

\item{y}{a list of population observations}

\item{model_config}{a list that contains parameters and network structure}
}
\value{
hiden states
}
\description{
update the agent states for the SIS model given observations such that the complete likelihood is not zero.
}
