% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_xx_gibbs_swap_alltimes.R
\name{sis_xx_gibbs_swap_alltimes}
\alias{sis_xx_gibbs_swap_alltimes}
\title{Gibbs sampler for SIS model}
\usage{
sis_xx_gibbs_swap_alltimes(xx, model_config)
}
\arguments{
\item{xx}{a binary matrix of size model_config$N by length(y), can be initialized from sis_init_hidden_popobservation.R}

\item{model_config}{a list containing parameters and network structure}
}
\value{
a new sample
}
\description{
update the agent states for the SIS model given observations such that the complete likelihood is not zero. 
It performs N swaps at each time step.
}
