% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_xx_swap_at_time_kernel.R
\name{sis_xx_gibbs_swap_at_time_kernel}
\alias{sis_xx_gibbs_swap_at_time_kernel}
\title{SIS swap kernel}
\usage{
sis_xx_gibbs_swap_at_time_kernel(xx, time_step, model_config)
}
\arguments{
\item{xx}{a matrix of size model_config$N by length(y), can be initialized from sis_init_hidden_popobservation.R}

\item{time_step}{0 <=t <= T}

\item{model_config}{a list that must contain
\itemize{
\item N : population size
\item rho : reporting rate
\item alpha0 : vector of probabilities
\item state_to_alpha: function to compute alpha given agent states at a previous time 
}}
}
\value{
xx
}
\description{
update the agent states for the SIS model given observations such that the complete likelihood is not zero.
}
