% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_xx_initialize.R
\name{sis_xx_initialize}
\alias{sis_xx_initialize}
\title{Initalize Gibbs sampler for SIS model}
\usage{
sis_xx_initialize(y, model_config)
}
\arguments{
\item{y}{a list of population observations}

\item{model_config}{a list that must contain
\itemize{
\item N : population size
\item rho : reporting rate
}}
}
\value{
agent_states
}
\description{
initialize the agent states for the SIS model given observations such that the complete likelihood is not zero.
}
