% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallpox_csmc.R
\name{smallpox_csmc}
\alias{smallpox_csmc}
\title{controlled SMC for the smallpox dataset}
\usage{
smallpox_csmc(y, model_config, logpolicy, particle_config)
}
\arguments{
\item{y}{, observations, length (T+1)}

\item{model_config}{a list containing model parameters, must pass the test of check_model_config;}

\item{logpolicy}{a 3-d array storing the bif information filter}

\item{particle_config}{a list specifying the particle filter, must pass the test of check_particle_config;}
}
\value{
the particle filter outcome, must contain estimate of log marginal likelihood and effective sample size.
}
\description{
marginal likelihood estimations and samples from the posterior distribution of agent states.
}
