% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallpox_csmc_fast.R
\name{smallpox_csmc_fast}
\alias{smallpox_csmc_fast}
\title{controlled SMC for the smallpox dataset.}
\usage{
smallpox_csmc_fast(
  y,
  model_config,
  logpolicy,
  num_particles = 20,
  ess_threshold = 0.5
)
}
\arguments{
\item{y}{, observations, length (T+1)}

\item{model_config}{a list containing model parameters, must pass the test of check_model_config;}

\item{logpolicy}{a 3-d array storing the bif information filter}

\item{num_particles}{number of particles, default to 20}

\item{ess_threshold}{adaptive resampling ESS threshold, default to 0.5}
}
\value{
estimated log marginal likelihood
}
\description{
marginal likelihood estimations and samples from the posterior distribution of agent states. 
This function is the same as smallpox_csmc, except that it only returns the log marginal likelihood.
}
