% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_conditional_pmf_i.R
\name{static_conditional_pmf_i}
\alias{static_conditional_pmf_i}
\title{Conditional PMF of I in the static model}
\usage{
static_conditional_pmf_i(y, alpha, rho, N)
}
\arguments{
\item{y}{observation (0 <= y <= N)}

\item{alpha}{a length N vector of probabilities}

\item{rho}{reporting rate}

\item{N}{length of alpha}
}
\value{
a list containing 
\itemize{
\item cpmf:conditional pmf of i given y as a vector of probabilities
\item llik:log-likelihood of y 
}
}
\description{
computes the marginal likelihood observing y given alpha in the static case and computes the conditional pmf of each i
}
