% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_get_alpha.R
\name{static_get_alpha}
\alias{static_get_alpha}
\title{Parametrization of the static model}
\usage{
static_get_alpha(beta, features)
}
\arguments{
\item{beta}{length is number of features}

\item{features}{N by num_features matrix}
}
\value{
the alpha vector
}
\description{
Parametrization of the static model
}
