% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_loglikelihood_marginal.R
\name{static_loglikelihood_marginal}
\alias{static_loglikelihood_marginal}
\title{Static Model Log Likelihood}
\usage{
static_loglikelihood_marginal(
  y,
  parameters,
  model_config,
  method = "tp",
  particle_config = NULL
)
}
\arguments{
\item{parameters}{a list containing beta and rho}

\item{model_config}{a list containing population size and agent features}

\item{method}{can be from the list c('tp','exact','mc')}

\item{particle_config}{if method == 'mc', then use this argument and it must contain num_particles}
}
\description{
compute log marginal likelihood of observing y in the static model
}
