% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/static_xx_swap.R
\name{static_xx_swap}
\alias{static_xx_swap}
\title{Swap move for Conditional Bernoulli target}
\usage{
static_xx_swap(xx, alpha)
}
\arguments{
\item{xx}{current sample}

\item{alpha}{vector of success probabilities}
}
\value{
xx_new
}
\description{
For a vector of length N, performs N swap moves. This algorithm runs in O(N) operations.
}
