% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resampling.R
\name{systematic_resampling}
\alias{systematic_resampling}
\title{Systematic resampling scheme}
\usage{
systematic_resampling(weights, ndraws)
}
\arguments{
\item{weights}{a vector of normalized weights}

\item{ndraws}{integer specifying number of ancestor indexes required}
}
\value{
a vector of ancestor indexes
}
\description{
Implements the systematic resampling scheme
}
