% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/under_reporting.R
\name{under_reporting_model}
\alias{under_reporting_model}
\title{Under reporting observation model}
\usage{
under_reporting_model(rho)
}
\arguments{
\item{rho}{under reporting parameter value}
}
\value{
a list containing: 
\itemize{
\item 'rmeasurement': function to sample from observation model
\item 'dmeasurement: function evaluating observation density
\item 'infected_support': function returning support of number of infected given observation  
}
}
\description{
Returns objects defining under reporting observation model
}
