% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boarding_csmc_full_fast.R
\name{boarding_csmc_full_fast}
\alias{boarding_csmc_full_fast}
\title{controlled SMC sampler for the boarding school dataset on complete networks.}
\usage{
boarding_csmc_full_fast(
  y,
  N,
  alpha0,
  lambda,
  gamma,
  rho,
  logpolicy,
  num_particles = 20,
  ess_threshold = 0.5
)
}
\arguments{
\item{y}{, observations, length (T+1)}

\item{N, }{population size}

\item{alpha0, }{double, initial infection probability}

\item{lambda, }{double, transmission potential}

\item{gamma, }{double, recovery rate}

\item{rho, }{double, reporting rate}

\item{logpolicy}{a 3-d array storing the bif information filter}

\item{num_particles}{number of particles, default to 20}

\item{ess_threshold}{adaptive resampling ESS threshold, default to 0.5}
}
\value{
estimated log marginal likelihood
}
\description{
marginal likelihood estimations.
}
