% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_config.R
\name{check_model_config}
\alias{check_model_config}
\title{check  model configurations}
\usage{
check_model_config(model_config)
}
\arguments{
\item{model_config}{a list that must contain: TODO}
}
\description{
this functionc checks if model_config is eligible as input for simulations, particle filters and gibbs samplers.
}
