% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logdsumbin.R
\name{logdsumbin}
\alias{logdsumbin}
\title{Sum of two independent Binomials}
\usage{
logdsumbin(it, lbar, gbar, N)
}
\arguments{
\item{it}{number of infections.}

\item{lbar}{average infection rate.}

\item{gbar}{average recovery rate.}

\item{N}{integer, must satisfy N = length(lbar).}
}
\value{
log transition densities p(i(t+1) = 0 : N | it, lbar, gbar);
}
\description{
computes the log density of SumBinomial distribition, parametrized by lbar, gbar and it.
}
