% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mylast.R
\name{mylast}
\alias{mylast}
\title{Access the last element in a vector}
\usage{
mylast(x, rm.na = FALSE)
}
\arguments{
\item{x}{vector}

\item{rm.na}{binary, if TRUE, then ignore the NA values in x.}
}
\value{
last element in the vector x
}
\description{
Access the last element in a vector
}
