% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_get_alpha.R
\name{sir_get_alpha_sir}
\alias{sir_get_alpha_sir}
\alias{sir_get_alpha_i}
\alias{sir_get_alpha_plusone}
\title{SIR model transition kernels}
\usage{
sir_get_alpha_sir(agent_states, model_config)

sir_get_alpha_i(agent_states, model_config)

sir_get_alpha_plusone(agent_states, model_config)
}
\arguments{
\item{agent_states}{the agent states at some time t}

\item{model_config}{a list containing model parameters}
}
\value{
\code{sir_get_alpha_sir(agent_states,model_config)} returns a matrix with three rows : alpha_s, alpha_i and alpha_r.

\code{sir_get_alpha_i(agent_states, model_config)} returns alpha_i

\code{sir_get_alpha_plusone(agent_states, model_config)} returns a vector, whose entries are alpha_i or alpha_r or 0 when previous states are 0, 1, 2 respectively.
}
\description{
the alpha function for SIR process.
}
