% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_kernel_twisted.R
\name{sir_kernel_twisted_ir}
\alias{sir_kernel_twisted_ir}
\title{Twisted SIR kernel}
\usage{
sir_kernel_twisted_ir(xxprev, it, rt, model_config)
}
\arguments{
\item{it}{}
}
\value{
a sample of xxprev
}
\description{
samples from xt given it, rt and x(t-1)
}
