% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sir_logdpoismulti_given_xxprev.R
\name{sir_logdpoismulti_given_xxprev}
\alias{sir_logdpoismulti_given_xxprev}
\title{SIR model}
\usage{
sir_logdpoismulti_given_xxprev(xxprev, model_config)
}
\arguments{
\item{xxprev}{x[t-1]}

\item{model_config}{a list containing parameters and network structure}
}
\description{
the density of p(it, rt | x[t-1])
}
