% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_bpf.R
\name{sis_bpf}
\alias{sis_bpf}
\title{Bootstrap particle filter for SIS model}
\usage{
sis_bpf(model_config, observations, filter_config)
}
\arguments{
\item{model_config}{a list containing:
\itemize{
\item 'N': size of the population
\item 'alpha0': vector of initial infection probability of length N
\item 'lambda': vector of infection rates of length N 
\item 'gamma': vector of recovery rates of length N 
\item 'network_type': character "full" or "net" specifying if network is fully connected or has a given network structure
\item 'adjacency': matrix indicating interaction of agents of dimension N x N (if model_config$network_type == "net)
\item 'degree' vector of node degree of length N (if model_config$network_type == "net)
\item 'dmeasurement': function evaluating observation density
}}

\item{observations}{a list containing:
\itemize{
\item 'T': number of observations is T+1
\item 'y': vector of observations of length T+1
}}

\item{filter_config}{a list containing:
\itemize{
\item 'P': number of particles
\item 'store': logical indicating if particle states and their ancestries are stored
\item 'resampling': character specifying type of resampling scheme
}}
}
\value{
a list containing: 
\itemize{
\item 'log_likelihood': vector of log-likelihood estimates of length T+1  
\item 'ess: vector of effective sample sizes of length T+1
\item 'runtime': algorithmic run time 
\item 'particles': array of particle states of dimension T+1 x N x P (if filter_config$store == TRUE)
\item 'ancestries': matrix of ancestor indexes of dimension T x P (if filter_config$store == TRUE)
}
}
\description{
Runs a bootstrap particle filter for SIS agent-based model
}
