% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sis_simulate_countinuous.R
\name{sis_simulate_continuous}
\alias{sis_simulate_continuous}
\title{SIS process}
\usage{
sis_simulate_continuous(agent_state, model_config, dt = 0.5)
}
\arguments{
\item{agent_state}{binary vector}

\item{model_config}{a list containing model parameters, must pass the test of check_model_config.}

\item{dt}{increment in time , dt < 1}
}
\description{
Markov kernel for SIS process when the increment in time dt < 1
}
